Category for conversation message to check if file is downloaded
