+++ Wire-iOS/Sources/UserInterface/Collections/CollectionAudioCell.swift
@@ -71,12 +71,7 @@ final public class CollectionAudioCell: CollectionCell {
     override open func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {
         switch action {
         case #selector(CollectionCell.forward(_:)):
-            if let _ = self.message?.fileMessageData?.fileURL {
-                return true
-            }
-            else {
-                return false
-            }
+            return self.message?.isFileDownloaded() ?? false
         default:
             return super.canPerformAction(action, withSender: sender)
         }
